requirements.txt
Django==2.2.10
djangorestframework==3.10.2
pytz==2019.2
sqlparse==0.3.0
autopep8==1.4.4
boto3==1.9.203
botocore==1.12.203
django-storages==1.7.1
djangorestframework==3.10.2
docutils==0.14
jmespath==0.9.4
Pillow==6.2.1
pycodestyle==2.5.0
python-dateutil==2.8.0
pytz==2019.2
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
markdown==3.0.0